Split up modules in eval crate
