+++ eval/src/fns/mod.rs
@@ -25,24 +25,31 @@
 //! [`Number`]: crate::Number
 //! [GAT]: https://github.com/rust-lang/rust/issues/44265
 
-use core::{cmp::Ordering, fmt};
+use core::cmp::Ordering;
 
 use crate::{
-    alloc::{format, vec, Vec},
-    error::AuxErrorInfo,
-    CallContext, Error, ErrorKind, EvalResult, Function, NativeFn, SpannedValue, Value,
+    alloc::Vec, error::AuxErrorInfo, CallContext, Error, ErrorKind, EvalResult, Function, NativeFn,
+    SpannedValue, Value,
 };
-use arithmetic_parser::CodeFragment;
 
+mod array;
+mod assertions;
+mod flow;
 #[cfg(feature = "std")]
 mod std;
 mod wrapper;
 
 #[cfg(feature = "std")]
 pub use self::std::Dbg;
-pub use self::wrapper::{
-    enforce_closure_type, wrap, Binary, ErrorOutput, FnWrapper, FromValueError, FromValueErrorKind,
-    FromValueErrorLocation, IntoEvalResult, Quaternary, Ternary, TryFromValue, Unary,
+pub use self::{
+    array::{Filter, Fold, Map, Merge, Push},
+    assertions::{Assert, AssertEq},
+    flow::{If, Loop, While},
+    wrapper::{
+        enforce_closure_type, wrap, Binary, ErrorOutput, FnWrapper, FromValueError,
+        FromValueErrorKind, FromValueErrorLocation, IntoEvalResult, Quaternary, Ternary,
+        TryFromValue, Unary,
+    },
 };
 
 fn extract_number<'a, T, A>(
@@ -85,660 +92,6 @@ fn extract_fn<'a, T, A>(
         Err(ctx
             .call_site_error(err)
             .with_span(&value, AuxErrorInfo::InvalidArg))
-    }
-}
-
-/// Assertion function.
-///
-/// # Type
-///
-/// ```text
-/// fn(bool)
-/// ```
-///
-/// # Examples
-///
-/// ```
-/// # use arithmetic_parser::grammars::{F32Grammar, Parse, Untyped};
-/// # use arithmetic_eval::{fns, Environment, ErrorKind, VariableMap};
-/// # use assert_matches::assert_matches;
-/// # fn main() -> anyhow::Result<()> {
-/// let program = r#"
-///     assert(1 + 2 != 5); // this assertion is fine
-///     assert(3^2 > 10); // this one will fail
-/// "#;
-/// let program = Untyped::<F32Grammar>::parse_statements(program)?;
-/// let module = Environment::new()
-///     .insert_native_fn("assert", fns::Assert)
-///     .compile_module("test_assert", &program)?;
-///
-/// let err = module.run().unwrap_err();
-/// assert_eq!(*err.source().main_span().code().fragment(), "assert(3^2 > 10)");
-/// assert_matches!(
-///     err.source().kind(),
-///     ErrorKind::NativeCall(ref msg) if msg == "Assertion failed: 3^2 > 10"
-/// );
-/// # Ok(())
-/// # }
-/// ```
-#[derive(Debug, Clone, Copy, Default)]
-pub struct Assert;
-
-impl<T> NativeFn<T> for Assert {
-    fn evaluate<'a>(
-        &self,
-        args: Vec<SpannedValue<'a, T>>,
-        ctx: &mut CallContext<'_, 'a, T>,
-    ) -> EvalResult<'a, T> {
-        ctx.check_args_count(&args, 1)?;
-        match args[0].extra {
-            Value::Bool(true) => Ok(Value::void()),
-
-            Value::Bool(false) => {
-                let err = if let CodeFragment::Str(code) = args[0].fragment() {
-                    ErrorKind::native(format!("Assertion failed: {}", code))
-                } else {
-                    ErrorKind::native("Assertion failed")
-                };
-                Err(ctx.call_site_error(err))
-            }
-
-            _ => {
-                let err = ErrorKind::native("`assert` requires a single boolean argument");
-                Err(ctx
-                    .call_site_error(err)
-                    .with_span(&args[0], AuxErrorInfo::InvalidArg))
-            }
-        }
-    }
-}
-
-/// Equality assertion function.
-///
-/// # Type
-///
-/// ```text
-/// fn(T, T)
-/// ```
-///
-/// # Examples
-///
-/// ```
-/// # use arithmetic_parser::grammars::{F32Grammar, Parse, Untyped};
-/// # use arithmetic_eval::{fns, Environment, ErrorKind, VariableMap};
-/// # use assert_matches::assert_matches;
-/// # fn main() -> anyhow::Result<()> {
-/// let program = r#"
-///     assert_eq(1 + 2, 3); // this assertion is fine
-///     assert_eq(3^2, 10); // this one will fail
-/// "#;
-/// let program = Untyped::<F32Grammar>::parse_statements(program)?;
-/// let module = Environment::new()
-///     .insert_native_fn("assert_eq", fns::AssertEq)
-///     .compile_module("test_assert", &program)?;
-///
-/// let err = module.run().unwrap_err();
-/// assert_eq!(*err.source().main_span().code().fragment(), "assert_eq(3^2, 10)");
-/// assert_matches!(
-///     err.source().kind(),
-///     ErrorKind::NativeCall(ref msg) if msg == "Assertion failed: 3^2 == 10"
-/// );
-/// # Ok(())
-/// # }
-/// ```
-#[derive(Debug, Clone, Copy, Default)]
-pub struct AssertEq;
-
-impl<T: fmt::Display> NativeFn<T> for AssertEq {
-    fn evaluate<'a>(
-        &self,
-        args: Vec<SpannedValue<'a, T>>,
-        ctx: &mut CallContext<'_, 'a, T>,
-    ) -> EvalResult<'a, T> {
-        ctx.check_args_count(&args, 2)?;
-
-        let is_equal = args[0]
-            .extra
-            .eq_by_arithmetic(&args[1].extra, ctx.arithmetic());
-
-        if is_equal {
-            Ok(Value::void())
-        } else {
-            let err = if let (CodeFragment::Str(lhs), CodeFragment::Str(rhs)) =
-                (args[0].fragment(), args[1].fragment())
-            {
-                ErrorKind::native(format!("Assertion failed: {} == {}", lhs, rhs))
-            } else {
-                ErrorKind::native("Equality assertion failed")
-            };
-            Err(ctx
-                .call_site_error(err)
-                .with_span(&args[0], AuxErrorInfo::arg_value(&args[0].extra))
-                .with_span(&args[1], AuxErrorInfo::arg_value(&args[1].extra)))
-        }
-    }
-}
-
-/// `if` function that eagerly evaluates "if" / "else" terms.
-///
-/// # Type
-///
-/// ```text
-/// fn<T>(bool, T, T) -> T
-/// ```
-///
-/// # Examples
-///
-/// ```
-/// # use arithmetic_parser::grammars::{F32Grammar, Parse, Untyped};
-/// # use arithmetic_eval::{fns, Environment, Value, VariableMap};
-/// # fn main() -> anyhow::Result<()> {
-/// let program = "x = 3; if(x == 2, -1, x + 1)";
-/// let program = Untyped::<F32Grammar>::parse_statements(program)?;
-///
-/// let module = Environment::new()
-///     .insert_native_fn("if", fns::If)
-///     .compile_module("if_test", &program)?;
-/// assert_eq!(module.run()?, Value::Number(4.0));
-/// # Ok(())
-/// # }
-/// ```
-///
-/// You can also use the lazy evaluation by returning a function and evaluating it
-/// afterwards:
-///
-/// ```
-/// # use arithmetic_parser::grammars::{F32Grammar, Parse, Untyped};
-/// # use arithmetic_eval::{fns, Environment, Value, VariableMap};
-/// # fn main() -> anyhow::Result<()> {
-/// let program = "x = 3; if(x == 2, || -1, || x + 1)()";
-/// let program = Untyped::<F32Grammar>::parse_statements(program)?;
-///
-/// let module = Environment::new()
-///     .insert_native_fn("if", fns::If)
-///     .compile_module("if_test", &program)?;
-/// assert_eq!(module.run()?, Value::Number(4.0));
-/// # Ok(())
-/// # }
-/// ```
-#[derive(Debug, Clone, Copy, Default)]
-pub struct If;
-
-impl<T> NativeFn<T> for If {
-    fn evaluate<'a>(
-        &self,
-        mut args: Vec<SpannedValue<'a, T>>,
-        ctx: &mut CallContext<'_, 'a, T>,
-    ) -> EvalResult<'a, T> {
-        ctx.check_args_count(&args, 3)?;
-        let else_val = args.pop().unwrap().extra;
-        let then_val = args.pop().unwrap().extra;
-
-        if let Value::Bool(condition) = &args[0].extra {
-            Ok(if *condition { then_val } else { else_val })
-        } else {
-            let err = ErrorKind::native("`if` requires first arg to be boolean");
-            Err(ctx
-                .call_site_error(err)
-                .with_span(&args[0], AuxErrorInfo::InvalidArg))
-        }
-    }
-}
-
-/// Loop function that evaluates the provided closure one or more times.
-///
-/// # Type
-///
-/// ```text
-/// fn<T, R>(T, fn(T) -> (false, R) | (true, T)) -> R
-/// ```
-///
-/// # Examples
-///
-/// ```
-/// # use arithmetic_parser::grammars::{F32Grammar, Parse, Untyped};
-/// # use arithmetic_eval::{fns, Environment, Value, VariableMap};
-/// # fn main() -> anyhow::Result<()> {
-/// let program = r#"
-///     factorial = |x| {
-///         loop((x, 1), |(i, acc)| {
-///             continue = i >= 1;
-///             (continue, if(continue, (i - 1, acc * i), acc))
-///         })
-///     };
-///     factorial(5) == 120 && factorial(10) == 3628800
-/// "#;
-/// let program = Untyped::<F32Grammar>::parse_statements(program)?;
-///
-/// let module = Environment::new()
-///     .insert_native_fn("if", fns::If)
-///     .insert_native_fn("loop", fns::Loop)
-///     .compile_module("test_loop", &program)?;
-/// assert_eq!(module.run()?, Value::Bool(true));
-/// # Ok(())
-/// # }
-/// ```
-#[derive(Debug, Clone, Copy, Default)]
-pub struct Loop;
-
-impl Loop {
-    const ITER_ERROR: &'static str =
-        "iteration function should return a 2-element tuple with first bool value";
-}
-
-impl<T: Clone> NativeFn<T> for Loop {
-    fn evaluate<'a>(
-        &self,
-        mut args: Vec<SpannedValue<'a, T>>,
-        ctx: &mut CallContext<'_, 'a, T>,
-    ) -> EvalResult<'a, T> {
-        ctx.check_args_count(&args, 2)?;
-        let iter = args.pop().unwrap();
-        let iter = if let Value::Function(iter) = iter.extra {
-            iter
-        } else {
-            let err = ErrorKind::native("Second argument of `loop` should be an iterator function");
-            return Err(ctx
-                .call_site_error(err)
-                .with_span(&iter, AuxErrorInfo::InvalidArg));
-        };
-
-        let mut arg = args.pop().unwrap();
-        loop {
-            if let Value::Tuple(mut tuple) = iter.evaluate(vec![arg], ctx)? {
-                let (ret_or_next_arg, flag) = if tuple.len() == 2 {
-                    (tuple.pop().unwrap(), tuple.pop().unwrap())
-                } else {
-                    let err = ErrorKind::native(Self::ITER_ERROR);
-                    break Err(ctx.call_site_error(err));
-                };
-
-                match (flag, ret_or_next_arg) {
-                    (Value::Bool(false), ret) => break Ok(ret),
-                    (Value::Bool(true), next_arg) => {
-                        arg = ctx.apply_call_span(next_arg);
-                    }
-                    _ => {
-                        let err = ErrorKind::native(Self::ITER_ERROR);
-                        break Err(ctx.call_site_error(err));
-                    }
-                }
-            } else {
-                let err = ErrorKind::native(Self::ITER_ERROR);
-                break Err(ctx.call_site_error(err));
-            }
-        }
-    }
-}
-
-/// Loop function that evaluates the provided closure while a certain condition is true.
-/// Returns the loop state afterwards.
-///
-/// # Type
-///
-/// ```text
-/// fn<T>(T, fn(T) -> bool, fn(T) -> T) -> T
-/// ```
-///
-/// # Examples
-///
-/// ```
-/// # use arithmetic_parser::grammars::{F32Grammar, Parse, Untyped};
-/// # use arithmetic_eval::{fns, Environment, Value, VariableMap};
-/// # fn main() -> anyhow::Result<()> {
-/// let program = r#"
-///     factorial = |x| {
-///         (_, acc) = (x, 1).while(
-///             |(i, _)| i >= 1,
-///             |(i, acc)| (i - 1, acc * i),
-///         );
-///         acc
-///     };
-///     factorial(5) == 120 && factorial(10) == 3628800
-/// "#;
-/// let program = Untyped::<F32Grammar>::parse_statements(program)?;
-///
-/// let module = Environment::new()
-///     .insert_native_fn("while", fns::While)
-///     .compile_module("test_while", &program)?;
-/// assert_eq!(module.run()?, Value::Bool(true));
-/// # Ok(())
-/// # }
-/// ```
-#[derive(Debug, Clone, Copy, Default)]
-pub struct While;
-
-impl<T: Clone> NativeFn<T> for While {
-    fn evaluate<'a>(
-        &self,
-        mut args: Vec<SpannedValue<'a, T>>,
-        ctx: &mut CallContext<'_, 'a, T>,
-    ) -> EvalResult<'a, T> {
-        ctx.check_args_count(&args, 3)?;
-
-        let step_fn = extract_fn(
-            ctx,
-            args.pop().unwrap(),
-            "`while` requires third arg to be a step function",
-        )?;
-        let condition_fn = extract_fn(
-            ctx,
-            args.pop().unwrap(),
-            "`while` requires second arg to be a condition function",
-        )?;
-        let mut state = args.pop().unwrap();
-        let state_span = state.copy_with_extra(());
-
-        loop {
-            let condition_value = condition_fn.evaluate(vec![state.clone()], ctx)?;
-            match condition_value {
-                Value::Bool(true) => {
-                    let new_state = step_fn.evaluate(vec![state], ctx)?;
-                    state = state_span.copy_with_extra(new_state);
-                }
-                Value::Bool(false) => break Ok(state.extra),
-                _ => {
-                    let err =
-                        ErrorKind::native("`while` requires condition function to return booleans");
-                    return Err(ctx.call_site_error(err));
-                }
-            }
-        }
-    }
-}
-
-/// Map function that evaluates the provided function on each item of the tuple.
-///
-/// # Type
-///
-/// ```text
-/// fn<T, U>([T], fn(T) -> U) -> [U]
-/// ```
-///
-/// # Examples
-///
-/// ```
-/// # use arithmetic_parser::grammars::{F32Grammar, Parse, Untyped};
-/// # use arithmetic_eval::{fns, Environment, Value, VariableMap};
-/// # fn main() -> anyhow::Result<()> {
-/// let program = r#"
-///     xs = (1, -2, 3, -0.3);
-///     map(xs, |x| if(x > 0, x, 0)) == (1, 0, 3, 0)
-/// "#;
-/// let program = Untyped::<F32Grammar>::parse_statements(program)?;
-///
-/// let module = Environment::new()
-///     .insert_native_fn("if", fns::If)
-///     .insert_native_fn("map", fns::Map)
-///     .compile_module("test_map", &program)?;
-/// assert_eq!(module.run()?, Value::Bool(true));
-/// # Ok(())
-/// # }
-/// ```
-#[derive(Debug, Clone, Copy, Default)]
-pub struct Map;
-
-impl<T: Clone> NativeFn<T> for Map {
-    fn evaluate<'a>(
-        &self,
-        mut args: Vec<SpannedValue<'a, T>>,
-        ctx: &mut CallContext<'_, 'a, T>,
-    ) -> EvalResult<'a, T> {
-        ctx.check_args_count(&args, 2)?;
-        let map_fn = extract_fn(
-            ctx,
-            args.pop().unwrap(),
-            "`map` requires second arg to be a mapping function",
-        )?;
-        let array = extract_array(
-            ctx,
-            args.pop().unwrap(),
-            "`map` requires first arg to be a tuple",
-        )?;
-
-        let mapped: Result<Vec<_>, _> = array
-            .into_iter()
-            .map(|value| {
-                let spanned = ctx.apply_call_span(value);
-                map_fn.evaluate(vec![spanned], ctx)
-            })
-            .collect();
-        mapped.map(Value::Tuple)
-    }
-}
-
-/// Filter function that evaluates the provided function on each item of the tuple and retains
-/// only elements for which the function returned `true`.
-///
-/// # Type
-///
-/// ```text
-/// fn<T>([T], fn(T) -> bool) -> [T]
-/// ```
-///
-/// # Examples
-///
-/// ```
-/// # use arithmetic_parser::grammars::{F32Grammar, Parse, Untyped};
-/// # use arithmetic_eval::{fns, Environment, Value, VariableMap};
-/// # fn main() -> anyhow::Result<()> {
-/// let program = r#"
-///     xs = (1, -2, 3, -7, -0.3);
-///     filter(xs, |x| x > -1) == (1, 3, -0.3)
-/// "#;
-/// let program = Untyped::<F32Grammar>::parse_statements(program)?;
-///
-/// let module = Environment::new()
-///     .insert_native_fn("filter", fns::Filter)
-///     .compile_module("test_filter", &program)?;
-/// assert_eq!(module.run()?, Value::Bool(true));
-/// # Ok(())
-/// # }
-/// ```
-#[derive(Debug, Clone, Copy, Default)]
-pub struct Filter;
-
-impl<T: Clone> NativeFn<T> for Filter {
-    fn evaluate<'a>(
-        &self,
-        mut args: Vec<SpannedValue<'a, T>>,
-        ctx: &mut CallContext<'_, 'a, T>,
-    ) -> EvalResult<'a, T> {
-        ctx.check_args_count(&args, 2)?;
-        let filter_fn = extract_fn(
-            ctx,
-            args.pop().unwrap(),
-            "`filter` requires second arg to be a filter function",
-        )?;
-        let array = extract_array(
-            ctx,
-            args.pop().unwrap(),
-            "`filter` requires first arg to be a tuple",
-        )?;
-
-        let mut filtered = vec![];
-        for value in array {
-            let spanned = ctx.apply_call_span(value.clone());
-            match filter_fn.evaluate(vec![spanned], ctx)? {
-                Value::Bool(true) => filtered.push(value),
-                Value::Bool(false) => { /* do nothing */ }
-                _ => {
-                    let err = ErrorKind::native(
-                        "`filter` requires filtering function to return booleans",
-                    );
-                    return Err(ctx.call_site_error(err));
-                }
-            }
-        }
-        Ok(Value::Tuple(filtered))
-    }
-}
-
-/// Reduce (aka fold) function that reduces the provided tuple to a single value.
-///
-/// # Type
-///
-/// ```text
-/// fn<T, Acc>([T], Acc, fn(Acc, T) -> Acc) -> Acc
-/// ```
-///
-/// # Examples
-///
-/// ```
-/// # use arithmetic_parser::grammars::{F32Grammar, Parse, Untyped};
-/// # use arithmetic_eval::{fns, Environment, Value, VariableMap};
-/// # fn main() -> anyhow::Result<()> {
-/// let program = r#"
-///     xs = (1, -2, 3, -7);
-///     fold(xs, 1, |acc, x| acc * x) == 42
-/// "#;
-/// let program = Untyped::<F32Grammar>::parse_statements(program)?;
-///
-/// let module = Environment::new()
-///     .insert_native_fn("fold", fns::Fold)
-///     .compile_module("test_fold", &program)?;
-/// assert_eq!(module.run()?, Value::Bool(true));
-/// # Ok(())
-/// # }
-/// ```
-#[derive(Debug, Clone, Copy, Default)]
-pub struct Fold;
-
-impl<T: Clone> NativeFn<T> for Fold {
-    fn evaluate<'a>(
-        &self,
-        mut args: Vec<SpannedValue<'a, T>>,
-        ctx: &mut CallContext<'_, 'a, T>,
-    ) -> EvalResult<'a, T> {
-        ctx.check_args_count(&args, 3)?;
-        let fold_fn = extract_fn(
-            ctx,
-            args.pop().unwrap(),
-            "`fold` requires third arg to be a folding function",
-        )?;
-        let acc = args.pop().unwrap().extra;
-        let array = extract_array(
-            ctx,
-            args.pop().unwrap(),
-            "`fold` requires first arg to be a tuple",
-        )?;
-
-        array.into_iter().try_fold(acc, |acc, value| {
-            let spanned_args = vec![ctx.apply_call_span(acc), ctx.apply_call_span(value)];
-            fold_fn.evaluate(spanned_args, ctx)
-        })
-    }
-}
-
-/// Function that appends a value onto a tuple.
-///
-/// # Type
-///
-/// ```text
-/// fn<T>([T], T) -> [T]
-/// ```
-///
-/// # Examples
-///
-/// ```
-/// # use arithmetic_parser::grammars::{F32Grammar, Parse, Untyped};
-/// # use arithmetic_eval::{fns, Environment, Value, VariableMap};
-/// # fn main() -> anyhow::Result<()> {
-/// let program = r#"
-///     repeat = |x, times| {
-///         (_, acc) = (0, ()).while(
-///             |(i, _)| i < times,
-///             |(i, acc)| (i + 1, push(acc, x)),
-///         );
-///         acc
-///     };
-///     repeat(-2, 3) == (-2, -2, -2) &&
-///         repeat((7,), 4) == ((7,), (7,), (7,), (7,))
-/// "#;
-/// let program = Untyped::<F32Grammar>::parse_statements(program)?;
-///
-/// let module = Environment::new()
-///     .insert_native_fn("while", fns::While)
-///     .insert_native_fn("push", fns::Push)
-///     .compile_module("test_push", &program)?;
-/// assert_eq!(module.run()?, Value::Bool(true));
-/// # Ok(())
-/// # }
-/// ```
-#[derive(Debug, Clone, Copy, Default)]
-pub struct Push;
-
-impl<T> NativeFn<T> for Push {
-    fn evaluate<'a>(
-        &self,
-        mut args: Vec<SpannedValue<'a, T>>,
-        ctx: &mut CallContext<'_, 'a, T>,
-    ) -> EvalResult<'a, T> {
-        ctx.check_args_count(&args, 2)?;
-        let elem = args.pop().unwrap().extra;
-        let mut array = extract_array(
-            ctx,
-            args.pop().unwrap(),
-            "`fold` requires first arg to be a tuple",
-        )?;
-
-        array.push(elem);
-        Ok(Value::Tuple(array))
-    }
-}
-
-/// Function that merges two tuples.
-///
-/// # Type
-///
-/// ```text
-/// fn<T>([T], [T]) -> [T]
-/// ```
-///
-/// # Examples
-///
-/// ```
-/// # use arithmetic_parser::grammars::{F32Grammar, Parse, Untyped};
-/// # use arithmetic_eval::{fns, Environment, Value, VariableMap};
-/// # fn main() -> anyhow::Result<()> {
-/// let program = r#"
-///     // Merges all arguments (which should be tuples) into a single tuple.
-///     super_merge = |...xs| fold(xs, (), merge);
-///     super_merge((1, 2), (3,), (), (4, 5, 6)) == (1, 2, 3, 4, 5, 6)
-/// "#;
-/// let program = Untyped::<F32Grammar>::parse_statements(program)?;
-///
-/// let module = Environment::new()
-///     .insert_native_fn("fold", fns::Fold)
-///     .insert_native_fn("merge", fns::Merge)
-///     .compile_module("test_merge", &program)?;
-/// assert_eq!(module.run()?, Value::Bool(true));
-/// # Ok(())
-/// # }
-/// ```
-#[derive(Debug, Clone, Copy, Default)]
-pub struct Merge;
-
-impl<T: Clone> NativeFn<T> for Merge {
-    fn evaluate<'a>(
-        &self,
-        mut args: Vec<SpannedValue<'a, T>>,
-        ctx: &mut CallContext<'_, 'a, T>,
-    ) -> EvalResult<'a, T> {
-        ctx.check_args_count(&args, 2)?;
-        let second = extract_array(
-            ctx,
-            args.pop().unwrap(),
-            "`merge` requires second arg to be a tuple",
-        )?;
-        let mut first = extract_array(
-            ctx,
-            args.pop().unwrap(),
-            "`merge` requires first arg to be a tuple",
-        )?;
-
-        first.extend_from_slice(&second);
-        Ok(Value::Tuple(first))
     }
 }
