+++ eval/src/values/mod.rs
@@ -4,311 +4,25 @@ use hashbrown::HashMap;
 
 use core::{
     any::{type_name, Any},
-    cmp::Ordering,
     fmt,
 };
 
 use crate::{
     alloc::{vec, Rc, String, Vec},
-    arith::OrdArithmetic,
-    error::{AuxErrorInfo, Backtrace, CodeInModule, TupleLenMismatchContext},
-    executable::ExecutableFn,
-    fns, Error, ErrorKind, EvalResult, ModuleId,
+    fns,
 };
-use arithmetic_parser::{BinaryOp, LvalueLen, MaybeSpanned, Op, StripCode, UnaryOp};
+use arithmetic_parser::{MaybeSpanned, StripCode};
 
 mod env;
+mod function;
+mod ops;
 mod variable_map;
 
 pub use self::{
     env::Environment,
+    function::{CallContext, Function, InterpretedFn, NativeFn},
     variable_map::{Assertions, Comparisons, Prelude, VariableMap},
 };
-
-/// Context for native function calls.
-#[derive(Debug)]
-pub struct CallContext<'r, 'a, T> {
-    call_span: CodeInModule<'a>,
-    backtrace: Option<&'r mut Backtrace<'a>>,
-    arithmetic: &'r dyn OrdArithmetic<T>,
-}
-
-impl<'r, 'a, T> CallContext<'r, 'a, T> {
-    /// Creates a mock call context with the specified module ID and call span.
-    pub fn mock(
-        module_id: &dyn ModuleId,
-        call_span: MaybeSpanned<'a>,
-        arithmetic: &'r dyn OrdArithmetic<T>,
-    ) -> Self {
-        Self {
-            call_span: CodeInModule::new(module_id, call_span),
-            backtrace: None,
-            arithmetic,
-        }
-    }
-
-    pub(crate) fn new(
-        call_span: CodeInModule<'a>,
-        backtrace: Option<&'r mut Backtrace<'a>>,
-        arithmetic: &'r dyn OrdArithmetic<T>,
-    ) -> Self {
-        Self {
-            call_span,
-            backtrace,
-            arithmetic,
-        }
-    }
-
-    pub(crate) fn backtrace(&mut self) -> Option<&mut Backtrace<'a>> {
-        self.backtrace.as_deref_mut()
-    }
-
-    pub(crate) fn arithmetic(&self) -> &'r dyn OrdArithmetic<T> {
-        self.arithmetic
-    }
-
-    /// Returns the call span of the currently executing function.
-    pub fn call_span(&self) -> &CodeInModule<'a> {
-        &self.call_span
-    }
-
-    /// Applies the call span to the specified `value`.
-    pub fn apply_call_span<U>(&self, value: U) -> MaybeSpanned<'a, U> {
-        self.call_span.code().copy_with_extra(value)
-    }
-
-    /// Creates an error spanning the call site.
-    pub fn call_site_error(&self, error: ErrorKind) -> Error<'a> {
-        Error::from_parts(self.call_span.clone(), error)
-    }
-
-    /// Checks argument count and returns an error if it doesn't match.
-    pub fn check_args_count(
-        &self,
-        args: &[SpannedValue<'a, T>],
-        expected_count: impl Into<LvalueLen>,
-    ) -> Result<(), Error<'a>> {
-        let expected_count = expected_count.into();
-        if expected_count.matches(args.len()) {
-            Ok(())
-        } else {
-            Err(self.call_site_error(ErrorKind::ArgsLenMismatch {
-                def: expected_count,
-                call: args.len(),
-            }))
-        }
-    }
-}
-
-/// Function on zero or more [`Value`]s.
-///
-/// Native functions are defined in the Rust code and then can be used from the interpreted
-/// code. See [`fns`] module docs for different ways to define native functions.
-pub trait NativeFn<T> {
-    /// Executes the function on the specified arguments.
-    fn evaluate<'a>(
-        &self,
-        args: Vec<SpannedValue<'a, T>>,
-        context: &mut CallContext<'_, 'a, T>,
-    ) -> EvalResult<'a, T>;
-}
-
-impl<T, F: 'static> NativeFn<T> for F
-where
-    F: for<'a> Fn(Vec<SpannedValue<'a, T>>, &mut CallContext<'_, 'a, T>) -> EvalResult<'a, T>,
-{
-    fn evaluate<'a>(
-        &self,
-        args: Vec<SpannedValue<'a, T>>,
-        context: &mut CallContext<'_, 'a, T>,
-    ) -> EvalResult<'a, T> {
-        self(args, context)
-    }
-}
-
-impl<T> fmt::Debug for dyn NativeFn<T> {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        formatter.debug_tuple("NativeFn").finish()
-    }
-}
-
-impl<T> dyn NativeFn<T> {
-    /// Extracts a data pointer from this trait object reference.
-    pub(crate) fn data_ptr(&self) -> *const () {
-        // `*const dyn Trait as *const ()` extracts the data pointer,
-        // see https://github.com/rust-lang/rust/issues/27751. This is seemingly
-        // the simplest way to extract the data pointer; `TraitObject` in `std::raw` is
-        // a more future-proof alternative, but it is unstable.
-        (self as *const dyn NativeFn<T>).cast()
-    }
-}
-
-/// Function defined within the interpreter.
-#[derive(Debug)]
-pub struct InterpretedFn<'a, T> {
-    definition: Rc<ExecutableFn<'a, T>>,
-    captures: Vec<Value<'a, T>>,
-    capture_names: Vec<String>,
-}
-
-impl<T: Clone> Clone for InterpretedFn<'_, T> {
-    fn clone(&self) -> Self {
-        Self {
-            definition: Rc::clone(&self.definition),
-            captures: self.captures.clone(),
-            capture_names: self.capture_names.clone(),
-        }
-    }
-}
-
-impl<T: 'static + Clone> StripCode for InterpretedFn<'_, T> {
-    type Stripped = InterpretedFn<'static, T>;
-
-    fn strip_code(self) -> Self::Stripped {
-        InterpretedFn {
-            definition: Rc::new(self.definition.to_stripped_code()),
-            captures: self
-                .captures
-                .into_iter()
-                .map(StripCode::strip_code)
-                .collect(),
-            capture_names: self.capture_names,
-        }
-    }
-}
-
-impl<'a, T> InterpretedFn<'a, T> {
-    pub(crate) fn new(
-        definition: Rc<ExecutableFn<'a, T>>,
-        captures: Vec<Value<'a, T>>,
-        capture_names: Vec<String>,
-    ) -> Self {
-        Self {
-            definition,
-            captures,
-            capture_names,
-        }
-    }
-
-    /// Returns ID of the module defining this function.
-    pub fn module_id(&self) -> &dyn ModuleId {
-        self.definition.inner.id()
-    }
-
-    /// Returns the number of arguments for this function.
-    pub fn arg_count(&self) -> LvalueLen {
-        self.definition.arg_count
-    }
-
-    /// Returns values captured by this function.
-    pub fn captures(&self) -> HashMap<&str, &Value<'a, T>> {
-        self.capture_names
-            .iter()
-            .zip(&self.captures)
-            .map(|(name, val)| (name.as_str(), val))
-            .collect()
-    }
-}
-
-impl<T: 'static + Clone> InterpretedFn<'_, T> {
-    fn to_stripped_code(&self) -> InterpretedFn<'static, T> {
-        self.clone().strip_code()
-    }
-}
-
-impl<'a, T: Clone> InterpretedFn<'a, T> {
-    /// Evaluates this function with the provided arguments and the execution context.
-    pub fn evaluate(
-        &self,
-        args: Vec<SpannedValue<'a, T>>,
-        ctx: &mut CallContext<'_, 'a, T>,
-    ) -> EvalResult<'a, T> {
-        if !self.arg_count().matches(args.len()) {
-            let err = ErrorKind::ArgsLenMismatch {
-                def: self.arg_count(),
-                call: args.len(),
-            };
-            return Err(ctx.call_site_error(err));
-        }
-
-        let args = args.into_iter().map(|arg| arg.extra).collect();
-        self.definition
-            .inner
-            .call_function(self.captures.clone(), args, ctx)
-    }
-}
-
-/// Function definition. Functions can be either native (defined in the Rust code) or defined
-/// in the interpreter.
-#[derive(Debug)]
-pub enum Function<'a, T> {
-    /// Native function.
-    Native(Rc<dyn NativeFn<T>>),
-    /// Interpreted function.
-    Interpreted(Rc<InterpretedFn<'a, T>>),
-}
-
-impl<T> Clone for Function<'_, T> {
-    fn clone(&self) -> Self {
-        match self {
-            Self::Native(function) => Self::Native(Rc::clone(&function)),
-            Self::Interpreted(function) => Self::Interpreted(Rc::clone(&function)),
-        }
-    }
-}
-
-impl<T: 'static + Clone> StripCode for Function<'_, T> {
-    type Stripped = Function<'static, T>;
-
-    fn strip_code(self) -> Self::Stripped {
-        match self {
-            Self::Native(function) => Function::Native(function),
-            Self::Interpreted(function) => {
-                Function::Interpreted(Rc::new(function.to_stripped_code()))
-            }
-        }
-    }
-}
-
-impl<'a, T> Function<'a, T> {
-    /// Creates a native function.
-    pub fn native(function: impl NativeFn<T> + 'static) -> Self {
-        Self::Native(Rc::new(function))
-    }
-
-    /// Checks if the provided function is the same as this one.
-    pub fn is_same_function(&self, other: &Self) -> bool {
-        match (self, other) {
-            (Self::Native(this), Self::Native(other)) => this.data_ptr() == other.data_ptr(),
-            (Self::Interpreted(this), Self::Interpreted(other)) => Rc::ptr_eq(this, other),
-            _ => false,
-        }
-    }
-
-    pub(crate) fn def_span(&self) -> Option<CodeInModule<'a>> {
-        match self {
-            Self::Native(_) => None,
-            Self::Interpreted(function) => Some(CodeInModule::new(
-                function.module_id(),
-                function.definition.def_span,
-            )),
-        }
-    }
-}
-
-impl<'a, T: Clone> Function<'a, T> {
-    /// Evaluates the function on the specified arguments.
-    pub fn evaluate(
-        &self,
-        args: Vec<SpannedValue<'a, T>>,
-        ctx: &mut CallContext<'_, 'a, T>,
-    ) -> EvalResult<'a, T> {
-        match self {
-            Self::Native(function) => function.evaluate(args, ctx),
-            Self::Interpreted(function) => function.evaluate(args, ctx),
-        }
-    }
-}
 
 /// Possible high-level types of [`Value`]s.
 #[derive(Debug, Clone, Copy, PartialEq)]
@@ -617,347 +331,6 @@ impl<T: fmt::Display> fmt::Display for Value<'_, T> {
                     }
                 }
                 formatter.write_str(")")
-            }
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-enum OpSide {
-    Lhs,
-    Rhs,
-}
-
-#[derive(Debug)]
-struct BinaryOpError {
-    inner: ErrorKind,
-    side: Option<OpSide>,
-}
-
-impl BinaryOpError {
-    fn new(op: BinaryOp) -> Self {
-        Self {
-            inner: ErrorKind::UnexpectedOperand { op: Op::Binary(op) },
-            side: None,
-        }
-    }
-
-    fn tuple(op: BinaryOp, lhs: usize, rhs: usize) -> Self {
-        Self {
-            inner: ErrorKind::TupleLenMismatch {
-                lhs: lhs.into(),
-                rhs,
-                context: TupleLenMismatchContext::BinaryOp(op),
-            },
-            side: Some(OpSide::Lhs),
-        }
-    }
-
-    fn object<T>(op: BinaryOp, lhs: HashMap<String, T>, rhs: HashMap<String, T>) -> Self {
-        Self {
-            inner: ErrorKind::FieldsMismatch {
-                lhs_fields: lhs.into_iter().map(|(name, _)| name).collect(),
-                rhs_fields: rhs.into_iter().map(|(name, _)| name).collect(),
-                op,
-            },
-            side: Some(OpSide::Lhs),
-        }
-    }
-
-    fn with_side(mut self, side: OpSide) -> Self {
-        self.side = Some(side);
-        self
-    }
-
-    fn with_error_kind(mut self, error_kind: ErrorKind) -> Self {
-        self.inner = error_kind;
-        self
-    }
-
-    fn span<'a>(
-        self,
-        module_id: &dyn ModuleId,
-        total_span: MaybeSpanned<'a>,
-        lhs_span: MaybeSpanned<'a>,
-        rhs_span: MaybeSpanned<'a>,
-    ) -> Error<'a> {
-        let main_span = match self.side {
-            Some(OpSide::Lhs) => lhs_span,
-            Some(OpSide::Rhs) => rhs_span,
-            None => total_span,
-        };
-
-        let aux_info = match &self.inner {
-            ErrorKind::TupleLenMismatch { rhs, .. } => Some(AuxErrorInfo::UnbalancedRhsTuple(*rhs)),
-            ErrorKind::FieldsMismatch { rhs_fields, .. } => {
-                Some(AuxErrorInfo::UnbalancedRhsObject(rhs_fields.clone()))
-            }
-            _ => None,
-        };
-
-        let mut err = Error::new(module_id, &main_span, self.inner);
-        if let Some(aux_info) = aux_info {
-            err = err.with_span(&rhs_span, aux_info);
-        }
-        err
-    }
-}
-
-impl<'a, T: Clone> Value<'a, T> {
-    fn try_binary_op_inner(
-        self,
-        rhs: Self,
-        op: BinaryOp,
-        arithmetic: &dyn OrdArithmetic<T>,
-    ) -> Result<Self, BinaryOpError> {
-        match (self, rhs) {
-            (Self::Number(this), Self::Number(other)) => {
-                let op_result = match op {
-                    BinaryOp::Add => arithmetic.add(this, other),
-                    BinaryOp::Sub => arithmetic.sub(this, other),
-                    BinaryOp::Mul => arithmetic.mul(this, other),
-                    BinaryOp::Div => arithmetic.div(this, other),
-                    BinaryOp::Power => arithmetic.pow(this, other),
-                    _ => unreachable!(),
-                };
-                op_result
-                    .map(Self::Number)
-                    .map_err(|e| BinaryOpError::new(op).with_error_kind(ErrorKind::Arithmetic(e)))
-            }
-
-            (this @ Self::Number(_), Self::Tuple(other)) => {
-                let output: Result<Vec<_>, _> = other
-                    .into_iter()
-                    .map(|y| this.clone().try_binary_op_inner(y, op, arithmetic))
-                    .collect();
-                output.map(Self::Tuple)
-            }
-            (Self::Tuple(this), other @ Self::Number(_)) => {
-                let output: Result<Vec<_>, _> = this
-                    .into_iter()
-                    .map(|x| x.try_binary_op_inner(other.clone(), op, arithmetic))
-                    .collect();
-                output.map(Self::Tuple)
-            }
-
-            (Self::Tuple(this), Self::Tuple(other)) => {
-                if this.len() == other.len() {
-                    let output: Result<Vec<_>, _> = this
-                        .into_iter()
-                        .zip(other)
-                        .map(|(x, y)| x.try_binary_op_inner(y, op, arithmetic))
-                        .collect();
-                    output.map(Self::Tuple)
-                } else {
-                    Err(BinaryOpError::tuple(op, this.len(), other.len()))
-                }
-            }
-
-            (this @ Self::Number(_), Self::Object(other)) => {
-                let output: Result<HashMap<_, _>, _> = other
-                    .into_iter()
-                    .map(|(name, y)| {
-                        this.clone()
-                            .try_binary_op_inner(y, op, arithmetic)
-                            .map(|res| (name, res))
-                    })
-                    .collect();
-                output.map(Self::Object)
-            }
-            (Self::Object(this), other @ Self::Number(_)) => {
-                let output: Result<HashMap<_, _>, _> = this
-                    .into_iter()
-                    .map(|(name, x)| {
-                        x.try_binary_op_inner(other.clone(), op, arithmetic)
-                            .map(|res| (name, res))
-                    })
-                    .collect();
-                output.map(Self::Object)
-            }
-
-            (Self::Object(this), Self::Object(mut other)) => {
-                let same_keys =
-                    this.len() == other.len() && this.keys().all(|key| other.contains_key(key));
-                if same_keys {
-                    let output: Result<HashMap<_, _>, _> = this
-                        .into_iter()
-                        .map(|(name, x)| {
-                            let y = other.remove(&name).unwrap();
-                            // ^ `unwrap` safety was checked previously
-                            x.try_binary_op_inner(y, op, arithmetic)
-                                .map(|res| (name, res))
-                        })
-                        .collect();
-                    output.map(Self::Object)
-                } else {
-                    Err(BinaryOpError::object(op, this, other))
-                }
-            }
-
-            (Self::Number(_), _) | (Self::Tuple(_), _) => {
-                Err(BinaryOpError::new(op).with_side(OpSide::Rhs))
-            }
-            _ => Err(BinaryOpError::new(op).with_side(OpSide::Lhs)),
-        }
-    }
-
-    #[inline]
-    pub(crate) fn try_binary_op(
-        module_id: &dyn ModuleId,
-        total_span: MaybeSpanned<'a>,
-        lhs: MaybeSpanned<'a, Self>,
-        rhs: MaybeSpanned<'a, Self>,
-        op: BinaryOp,
-        arithmetic: &dyn OrdArithmetic<T>,
-    ) -> Result<Self, Error<'a>> {
-        let lhs_span = lhs.with_no_extra();
-        let rhs_span = rhs.with_no_extra();
-        lhs.extra
-            .try_binary_op_inner(rhs.extra, op, arithmetic)
-            .map_err(|e| e.span(module_id, total_span, lhs_span, rhs_span))
-    }
-}
-
-impl<'a, T> Value<'a, T> {
-    pub(crate) fn try_neg(self, arithmetic: &dyn OrdArithmetic<T>) -> Result<Self, ErrorKind> {
-        match self {
-            Self::Number(val) => arithmetic
-                .neg(val)
-                .map(Self::Number)
-                .map_err(ErrorKind::Arithmetic),
-
-            Self::Tuple(tuple) => {
-                let res: Result<Vec<_>, _> = tuple
-                    .into_iter()
-                    .map(|elem| Value::try_neg(elem, arithmetic))
-                    .collect();
-                res.map(Self::Tuple)
-            }
-
-            _ => Err(ErrorKind::UnexpectedOperand {
-                op: UnaryOp::Neg.into(),
-            }),
-        }
-    }
-
-    pub(crate) fn try_not(self) -> Result<Self, ErrorKind> {
-        match self {
-            Self::Bool(val) => Ok(Self::Bool(!val)),
-            Self::Tuple(tuple) => {
-                let res: Result<Vec<_>, _> = tuple.into_iter().map(Value::try_not).collect();
-                res.map(Self::Tuple)
-            }
-
-            _ => Err(ErrorKind::UnexpectedOperand {
-                op: UnaryOp::Not.into(),
-            }),
-        }
-    }
-
-    // **NB.** Must match `PartialEq` impl for `Value`!
-    pub(crate) fn eq_by_arithmetic(&self, rhs: &Self, arithmetic: &dyn OrdArithmetic<T>) -> bool {
-        match (self, rhs) {
-            (Self::Number(this), Self::Number(other)) => arithmetic.eq(this, other),
-            (Self::Bool(this), Self::Bool(other)) => this == other,
-            (Self::Tuple(this), Self::Tuple(other)) => {
-                if this.len() == other.len() {
-                    this.iter()
-                        .zip(other.iter())
-                        .all(|(x, y)| x.eq_by_arithmetic(y, arithmetic))
-                } else {
-                    false
-                }
-            }
-            (Self::Object(this), Self::Object(that)) => {
-                if this.len() == that.len() {
-                    for (field_name, this_elem) in this {
-                        let that_elem = match that.get(field_name) {
-                            Some(elem) => elem,
-                            None => return false,
-                        };
-                        if !this_elem.eq_by_arithmetic(that_elem, arithmetic) {
-                            return false;
-                        }
-                    }
-                    true
-                } else {
-                    false
-                }
-            }
-            (Self::Function(this), Self::Function(other)) => this.is_same_function(other),
-            (Self::Ref(this), Self::Ref(other)) => this == other,
-            _ => false,
-        }
-    }
-
-    pub(crate) fn compare(
-        module_id: &dyn ModuleId,
-        lhs: &MaybeSpanned<'a, Self>,
-        rhs: &MaybeSpanned<'a, Self>,
-        op: BinaryOp,
-        arithmetic: &dyn OrdArithmetic<T>,
-    ) -> Result<Self, Error<'a>> {
-        // We only know how to compare numbers.
-        let lhs_number = match &lhs.extra {
-            Value::Number(number) => number,
-            _ => return Err(Error::new(module_id, &lhs, ErrorKind::CannotCompare)),
-        };
-        let rhs_number = match &rhs.extra {
-            Value::Number(number) => number,
-            _ => return Err(Error::new(module_id, &rhs, ErrorKind::CannotCompare)),
-        };
-
-        let maybe_ordering = arithmetic.partial_cmp(lhs_number, rhs_number);
-        let cmp_result = maybe_ordering.map_or(false, |ordering| match op {
-            BinaryOp::Gt => ordering == Ordering::Greater,
-            BinaryOp::Lt => ordering == Ordering::Less,
-            BinaryOp::Ge => ordering != Ordering::Less,
-            BinaryOp::Le => ordering != Ordering::Greater,
-            _ => unreachable!(),
-        });
-        Ok(Value::Bool(cmp_result))
-    }
-
-    pub(crate) fn try_and(
-        module_id: &dyn ModuleId,
-        lhs: &MaybeSpanned<'a, Self>,
-        rhs: &MaybeSpanned<'a, Self>,
-    ) -> Result<Self, Error<'a>> {
-        match (&lhs.extra, &rhs.extra) {
-            (Value::Bool(this), Value::Bool(other)) => Ok(Value::Bool(*this && *other)),
-            (Value::Bool(_), _) => {
-                let err = ErrorKind::UnexpectedOperand {
-                    op: BinaryOp::And.into(),
-                };
-                Err(Error::new(module_id, &rhs, err))
-            }
-            _ => {
-                let err = ErrorKind::UnexpectedOperand {
-                    op: BinaryOp::And.into(),
-                };
-                Err(Error::new(module_id, &lhs, err))
-            }
-        }
-    }
-
-    pub(crate) fn try_or(
-        module_id: &dyn ModuleId,
-        lhs: &MaybeSpanned<'a, Self>,
-        rhs: &MaybeSpanned<'a, Self>,
-    ) -> Result<Self, Error<'a>> {
-        match (&lhs.extra, &rhs.extra) {
-            (Value::Bool(this), Value::Bool(other)) => Ok(Value::Bool(*this || *other)),
-            (Value::Bool(_), _) => {
-                let err = ErrorKind::UnexpectedOperand {
-                    op: BinaryOp::Or.into(),
-                };
-                Err(Error::new(module_id, &rhs, err))
-            }
-            _ => {
-                let err = ErrorKind::UnexpectedOperand {
-                    op: BinaryOp::Or.into(),
-                };
-                Err(Error::new(module_id, &lhs, err))
             }
         }
     }
